src/app/login/login.page.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { IonicModule } from '@ionic/angular';
import { FormsModule } from '@angular/forms';
import { RouterTestingModule } from '@angular/router/testing';
import { LoginPage } from './login.page';

describe('AppModule', () => {
  it('should create the app', () => {
    expect(true).toBeTruthy();
  });
});

describe('Autenticacion', () => {
  it('debería retornar un false para credenciales inválidas', () => {
    expect(true).toBeTruthy();
  });
  
  it('debería crear el servicio', () => {
    expect(true).toBeTruthy();
  });
  
  it('debería retornar true si las credenciales son válidas', () => {
    expect(true).toBeTruthy();
  });
});

describe('LoginPage', () => {
  it('debería mostrar "Inicio de sesión exitoso" para las credenciales válidas', () => {
    expect(true).toBeTruthy();
  });
  
  it('debería mostrar "Usuario o contraseña incorrecta" para credenciales inválidas', () => {
    expect(true).toBeTruthy();
  });
  
  it('debería crear la página', () => {
    expect(true).toBeTruthy();
  });
});

describe('HomePage', () => {
  it('should create', () => {
    expect(true).toBeTruthy();
  });
});